pyproject.toml
[project]
name = "101-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-cohere>=0.5.0",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-mistralai>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.5",
    "langsmith>=0.5.0",
    "litellm>=1.80.16",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
